+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Disabled models do not expose an address (#910)
 - Increase max page size default (#922)
 - Remove OrchestratorDataSampleSerializer and OrchestratorDataSampleUpdateSerializer and use orchestrator client directly (#911)
+- Remove OrchestratorMetricSerializer and use orchestrator client directly (#912)
 
 ### Fixed
 - Compute CP dates after updating related tasks (#855)